requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
numpy==1.24.3
spacy==3.7.2
aiohttp==3.8.5
PyNaCl==1.5.0
gunicorn==21.2.0